backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
elasticsearch==8.11.0

# Authentication and Security
cryptography>=41.0.0,<46.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# AI and ML libraries
tensorflow>=2.15.0,<2.16.0
scikit-learn>=1.3.0,<1.4.0
numpy>=1.24.0,<1.25.0
pandas>=2.1.0,<2.2.0
mediapipe>=0.10.0,<0.11.0
opencv-python>=4.8.0,<4.9.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# Payment processing
stripe==7.8.0

# Task queue
celery==5.3.4
celery[redis]==5.3.4

# API documentation
drf-spectacular==0.26.5

# Environment and configuration
python-decouple==3.8
django-environ==0.11.2

# File handling
django-storages==1.14.2
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Monitoring and logging
sentry-sdk==1.38.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
